Rename test helper
