+++ lib/firmwareTestHelper.js
